version 3.10.11
client: allow openstreetmap zoom 17
server: allow openstreetmap zoom 17
        faa updates

+++ decosects/purgeoldcharts.sh
@@ -60,6 +60,11 @@ function purgedatum
             mv "$file" "purged-$file"
             continue
         fi
+        if [ "${file:0:12}" == "datums/DDTPP" ]
+        then
+            mv "$file" "purged-$file"
+            continue
+        fi
         if [ "${file:0:12}" == "datums/fixes" ]
         then
             rm -rf "$file"
